+++ ExadelBonusPlus.Services.Models/Bonus.cs
@@ -1,5 +1,4 @@
-﻿using ExadelBonusPlus.Services.Models.Interfaces;
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace ExadelBonusPlus.Services.Models
