+++ ExadelBonusPlus.Services.Models/DTO/LocationDto.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ExadelBonusPlus.WebApi.Dtos
+namespace ExadelBonusPlus.Services.Models
 {
     public class LocationDto
     {
