+++ ExadelBonusPlus.Services.Models/Interfaces/IBonusRepository.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using ExadelBonusPlus.Services.Models;
 
 namespace ExadelBonusPlus.Services.Models
 {
