+++ ExadelBonusPlus.Services.Models/Interfaces/IBonusService.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using ExadelBonusPlus.Services.Models;
 
 namespace ExadelBonusPlus.Services.Models
 {
