+++ ExadelBonusPlus.Services.Models/Interfaces/IEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace ExadelBonusPlus.Services.Models.Interfaces
+﻿namespace ExadelBonusPlus.Services.Models
 {
     using System;
