+++ ExadelBonusPlus.Services.Models/Interfaces/IRepository.cs
@@ -1,8 +1,3 @@
-
-﻿using ExadelBonusPlus.Services.Models.Interfaces;
-using System;
-using System.Linq.Expressions;
-using System.Text;
 ﻿using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
