+++ ExadelBonusPlus.Services.Models/Interfaces/ITokenRefreshRepository.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
