+++ ExadelBonusPlus.Services.Models/Interfaces/ITokenRefreshService.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
-using ExadelBonusPlus.Services.Models;
 
-namespace ExadelBonusPlus.Services
+namespace ExadelBonusPlus.Services.Models
 {
     public interface ITokenRefreshService
     {
