+++ ExadelBonusPlus.Services.Models/MapperProfile.cs
@@ -18,7 +18,16 @@ namespace ExadelBonusPlus.Services.Models
                     opt => opt.MapFrom<UserResolver>())
                 .ReverseMap();
 
-            CreateMap<ApplicationRole, RoleDTO.RoleDTO>().ReverseMap();
+            CreateMap<ApplicationRole, RoleDTO>().ReverseMap();
+
+            //From model to dto
+            CreateMap<Vendor, VendorDto>()
+                .ForMember(vd => vd.Location, opts => opts.MapFrom(v => v.Location));
+            CreateMap<Location, LocationDto>();
+
+            //From dto to model 
+            CreateMap<VendorDto, Vendor>()
+                .ForMember(v => v.Location, opts => opts.MapFrom(vd => new Location { City = vd.Location.City, Address = vd.Location.Address, Country = vd.Location.Country, Latitude = vd.Location.Latitude, Longitude = vd.Location.Longitude }));
         }
     }
 }
